Improved time calculation
Now the time calculation is not using for loop but only 1 expression(line 23): (time required for first inhabitant + time required for last inhabitant) / 2 * freespace. This expression is more accurate and shows times that are very close to real time.

I have also added calculation for time required for 10, 50, 100, 250 and 500 inhabitants(line 16). These numbers can be changed but the lower the index in array the number has the lower the number must be. The top 3 are shown in the town hall window under the "Be filled in" time. If the capacity should not be filled with current happiness level the infinity character \u221E is shown.