Adapting the code to a modified ajax response
IkaEasy currently does not calculate correctly the resources needed for a building -- the red/green indicators next to the buildings are off.
The reason is that it does not take into account the completed researches.
The root cause was eventually tracked to the fact that Ikariam has apparently slightly modified
the returned JSON that is used to determine the research levels.
This change adapts the code to the new model.