Fixing the pirate point conversion dialog
Currently the the slider in the pirate point conversion dialog does not work when IkaEasy is enabled.
It appears that the issue occurs when an item with class 'expandable' is added to the left menu.
For the time being the 'expandable' class is removed from the additional left menu items that we are adding.
There is no real difference in functionality as a result, but the pirate point conversion dialog now works.